+++ app/src/app/app.routing.module.ts
@@ -17,8 +17,8 @@ const routes: Routes = [
     loadChildren: './pages/home/home.routing.module#HomeRoutingModule'
   },
   {
-    path: 'impress',
-    loadChildren: './pages/impress/impress.routing.module#ImpressRoutingModule'
+    path: 'imprint',
+    loadChildren: './pages/imprint/imprint.routing.module#ImpressRoutingModule'
   },
   {
     path: 'data-protection',
