+++ app/src/app/core/components/footer/footer.component.html
@@ -2,7 +2,7 @@
   <div class="container">
     <div class="d-flex flex-wrap justify-content-center">
       <a class="p-3" [routerLink]="['/about']" i18n="Menu About item label@@menuAbout">About</a>
-      <a class="p-3" [routerLink]="['/impress']" i18n="Menu Legal Notice item label@@menuLegalNotice">Legal Notice</a>
+      <a class="p-3" [routerLink]="['/imprint']" i18n="Menu Legal Notice item label@@menuLegalNotice">Legal Notice</a>
       <a class="p-3" [routerLink]="['/data-protection']" i18n="Menu Data Protection item label@@menuDataProtection">Data Protection Declaration</a>
     </div>
   </div>
