+++ app/src/app/core/components/header/header.component.html
@@ -37,7 +37,7 @@
           <div class="dropdown-divider"></div>
           <a class="dropdown-item" [routerLink]="['/about']" i18n="Menu About item label@@menuAbout">About</a>
           <div class="dropdown-divider"></div>
-          <a class="dropdown-item" [routerLink]="['/impress']" i18n="Menu Legal Notice item label@@menuLegalNotice">Legal Notice</a>
+          <a class="dropdown-item" [routerLink]="['/imprint']" i18n="Menu Legal Notice item label@@menuLegalNotice">Legal Notice</a>
           <div class="dropdown-divider"></div>
           <a class="dropdown-item" [routerLink]="['/data-protection']" i18n="Menu Data Protection Declaration item label@@menuDataProtection"
             >Data Protection Declaration</a
