+++ app/src/app/pages/imprint/imprint.routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { SharedModule } from '../../shared/shared.module';
-import { ImpressComponent } from './impress.component';
+import { ImpressComponent } from './imprint.component';
 
 const impressRoutes: Routes = [
   {
